requirements.txt
Flask
qrcode
pillow
openpyxl
gunicorn
psycopg2-binary
pytz